% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col2hex.R
\name{col2hex}
\alias{col2hex}
\title{Convert color to hex}
\usage{
col2hex(color, alpha = 1)
}
\arguments{
\item{color}{Character string. An R color or color as hex.}

\item{alpha}{Transparency. A value from 0 to 1 (default).}
}
\value{
A character string of a hexadecimal color
}
\description{
Convert color to hex
}
\details{
This function is primarily to be used on R colors, and will convert
  them to hexmode sufficient for typical use in visualization, web
  presentation, etc.  If you already have a hex value, it can be useful for
  tacking on the alpha level.
}
\examples{
library(visibly)
col2hex('red')
col2hex('#ff5500', alpha = .1)
}
\seealso{
\link[base]{as.hexmode} \link[grDevices]{colors}
\link[grDevices]{col2rgb} \link[grDevices]{rgb}
}
