% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col2lab.R
\name{col2lab}
\alias{col2lab}
\title{Convert color to lab}
\usage{
col2lab(color)
}
\arguments{
\item{color}{An R color or a hexadecimal string of the form "#rrggbb"}
}
\value{
A matrix of LAB values useful for colorgorical starting points.
}
\description{
Converts an R color or a hexadecimal string to CIE Lab.
}
\details{
See the help for \link[grDevices]{convertColor}.
}
\examples{
library(visibly)
col2lab('red')
x = list('red', 'white', 'blue')
col2lab(x)

}
\seealso{
\link[visibly]{colorgorical} \link[grDevices]{col2rgb} \link[grDevices]{convertColor}
}
