% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_contrast_checker.R
\name{color_contrast_checker}
\alias{color_contrast_checker}
\title{Color contrast}
\usage{
color_contrast_checker(foreground, background = "#FFFFFF")
}
\arguments{
\item{foreground}{Character string. An R color or color as hex.}

\item{background}{Character string. An R color or color as hex. Default is
white/'#FFFFFF'.}
}
\value{
If you have \link[jsonlite]{jsonlite}, it will return a data frame of
  the ratio and which standards are passed. Otherwise you get the same
  information but as a character vector.
}
\description{
Get the color contrast of two colors
}
\details{
This uses the api at
  \href{https://webaim.org/resources/contrastchecker/}{webaim.org} to
  determine the contrast ratio of the foreground color to the background
  color.  This is useful to ensure chosen colors meet web standards for
  accessibility.
}
\examples{
\dontrun{
library(visibly)
color_contrast_checker(foreground = 'red', background = 'papayawhip')

# standard blue is not good for contrasting links from other text
color_contrast_checker(foreground = '#000080', background = 'black')
}
}
