% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_heat.R
\name{corr_heat}
\alias{corr_heat}
\title{Visualize a correlation matrix}
\usage{
corr_heat(cormat, n_factors = NULL, psychOptions = NULL,
  ordering = c("fa", "polar", "raw", "first", "absolute"), three_d = FALSE,
  diagonal = 1, dir = -1, pal = "vik")
}
\arguments{
\item{cormat}{A correlation matrix}

\item{n_factors}{The number of factors for the factor analysis. Default is NULL.}

\item{psychOptions}{a list of arguments to pass to \code{psych::fa}}

\item{ordering}{Order of the result. See details}

\item{three_d}{Return a surface plot.}

\item{diagonal}{Value for the diagonal. Must be 1 (default) or NA.}
}
\value{
A plotly object
}
\description{
Visualize a correlation matrix
}
\details{
Correlation matrices are typically better visualized rather than parsed
  numerically, and while one can do so with various packages at this point,
  they either don't reorder the data, don't show the actual values, or only
  order based on cluster analysis, and one often may not want the cluster
  based approach to ordering if dealing with a correlation matrix, which may
  be too small column-wise to be useful for a cluster analysis, or may be a
  specific type of data more amenable to a measurement error approach (e.g.
  items from a particular scale).

  \code{corrheat} produces a color coded matrix in which Blue represents
  positive, and Red, negative correlations, and fades to white the smaller
  the values are. It is interactive, such that one can hover over the image
  to obtain the correlation value.

  The ordering is based on the results of a factor analysis
  from the \link[psych]{psych} package (which is required). It can be based
  on \code{\link[psych]{fa.sort}} (default), max raw ("raw") loading,
  absolute ("absolute") values of the loadings, or polar coordinates (see
  \link[psych]{polar} ) across all factors.

  By default the number of factors is chosen to more likely 'just work' for
  visualization purposes.  If the number of columns is three or fewer, it
  will force only one factor, otherwise it will be
  \code{floor(sqrt(ncol(x)))}).  While you may supply the number of factors,
  as well as other options to the \link[psych]{fa} function via
  \code{psychOptions}, if you want explore a factor analysis you should you
  probably should do that separately.
}
\examples{
library(visibly)

corr_heat(cormat = cor(mtcars), ordering = 'polar')

}
