% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrheat.R
\name{corrheat}
\alias{corrheat}
\title{Title}
\usage{
corrheat(cormat, n_factors = NULL, psychOptions = NULL, ordering = c("fa",
  "raw", "first", "absolute"), three_d = FALSE, dir = -1)
}
\arguments{
\item{dir}{}
}
\description{
Title
}
