% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{palettes}
\alias{palettes}
\title{Commonly used palettes}
\format{(Possibly) named vectors/lists of colors

- orange : #ff5500

- stanred: #b2001d

- RdBu_11: RdBu palette from colorbrewer for sequence of 11

- Rblue: #1f65b7

- Latvian Red: #9E1B34

- Tyrian Purple: #3c6602

- Tyrian Purple2: #990024  A more web friendly version

- Plotly: plotly's default colorscale, because decent documentation is next to impossible for those folks.
...}
\source{
For example the colortools and colorspace packages, \url{http://paletton.com/}, \url{https://www.sessions.edu/color-calculator/},  \url{https://colorbrewer.org}
}
\usage{
palettes
}
\description{
A list of commonly used palettes with some base color and including complementary, monochromatic, analagous, split_complementary, triadic, tetradic.  Otherwise, just ones I use e.g. colorbrewer, plotly or others of interest.
}
\references{
https://stackoverflow.com/questions/40673490/how-to-get-plotly-js-default-colors-list
}
\keyword{datasets}
