% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coefficients.brmsfit.R
\name{plot_coefficients.brmsfit}
\alias{plot_coefficients.brmsfit}
\title{Plot fixed or random effects coefficients for brmsfit objects.}
\usage{
\method{plot_coefficients}{brmsfit}(model, order = "decreasing",
  sd_multi = 2, keep_intercept = FALSE, palette = "bilbao",
  ref_line = 0, trans = NULL, plot = TRUE, ranef = FALSE,
  which_ranef = NULL, ...)
}
\arguments{
\item{model}{The model that is the point of this function. For example, lm, glm, gam, lme4, brms.}

\item{order}{The order of the plots- "increasing", "decreasing", or a numeric
vector giving the order. The default is NULL, i.e. the default ordering. Not applied to random effects.}

\item{sd_multi}{The multiplier that determines the width of the interval. Default is 2.}

\item{keep_intercept}{Default is FALSE. Intercepts are typically on a very
different scale than covariate effects.}

\item{palette}{A scico palette. Default is 'bilbao'.}

\item{ref_line}{A reference line. Default is zero.}

\item{trans}{A transformation function to be applied to the coefficients
(e.g. exponentiation).}

\item{plot}{Default is TRUE, but sometimes you just want the data.}

\item{ranef}{If applicable, whether to plot random effects instead of fixed effects.}

\item{which_ranef}{If plotting random effects, which one to plot.}

\item{...}{Other arguments applied for specific methods.}
}
\value{
A ggplot of the coefficients and their interval estimates. Or the
  data that would be used to create the plot.
}
\description{
Plot fixed or random effects coefficients for brmsfit objects.
}
\examples{
# placeholder

}
\seealso{
Other model visualization: \code{\link{plot_coefficients.lm}},
  \code{\link{plot_coefficients.merMod}},
  \code{\link{plot_coefficients}},
  \code{\link{plot_gam_2d}}, \code{\link{plot_gam_3d}},
  \code{\link{plot_gam_check}}, \code{\link{plot_gam}}
}
\concept{model visualization}
