% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coefficients.lm.R
\name{plot_coefficients.lm}
\alias{plot_coefficients.lm}
\alias{plot_coefficients.glm}
\title{Plot coefficients for standard linear models.}
\usage{
\method{plot_coefficients}{lm}(model, order = "decreasing",
  sd_multi = 2, keep_intercept = FALSE, palette = "bilbao",
  ref_line = 0, trans = NULL, plot = TRUE, ...)

\method{plot_coefficients}{glm}(model, order = "decreasing",
  sd_multi = 2, keep_intercept = FALSE, palette = "bilbao",
  ref_line = 0, trans = NULL, plot = TRUE, ...)
}
\arguments{
\item{model}{The lm or glm model}

\item{order}{The order of the plots- "increasing", "decreasing", or a numeric
vector giving the order. The default is NULL, i.e. the default ordering. Not applied to random effects.}

\item{sd_multi}{The multiplier that determines the width of the interval. Default is 2.}

\item{keep_intercept}{Default is FALSE. Intercepts are typically on a very
different scale than covariate effects.}

\item{palette}{A scico palette. Default is 'bilbao'.}

\item{ref_line}{A reference line. Default is zero.}

\item{trans}{A transformation function to be applied to the coefficients
(e.g. exponentiation).}

\item{plot}{Default is TRUE, but sometimes you just want the data.}

\item{...}{Other arguments applied for specific methods.}
}
\value{
A ggplot of the coefficients and their interval estimates. Or the
  data that would be used to create the plot.
}
\description{
A basic plot of coefficients with their uncertainty interval for
  lm and glm objects.
}
\details{
This is more or less a function that serves as the basis for other models I actually use.
}
\examples{
mod = lm(mpg ~ ., mtcars)
plot_coefficients(mod, order = 'increasing')

}
\seealso{
Other model visualization: \code{\link{plot_coefficients.brmsfit}},
  \code{\link{plot_coefficients.merMod}},
  \code{\link{plot_coefficients}},
  \code{\link{plot_gam_2d}}, \code{\link{plot_gam_3d}},
  \code{\link{plot_gam_check}}, \code{\link{plot_gam}}

Other model visualization: \code{\link{plot_coefficients.brmsfit}},
  \code{\link{plot_coefficients.merMod}},
  \code{\link{plot_coefficients}},
  \code{\link{plot_gam_2d}}, \code{\link{plot_gam_3d}},
  \code{\link{plot_gam_check}}, \code{\link{plot_gam}}
}
\concept{model visualization}
