% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coefficients.merMod.R
\name{plot_coefficients.merMod}
\alias{plot_coefficients.merMod}
\title{Plot fixed or random effects coefficients for merMod objects.}
\usage{
\method{plot_coefficients}{merMod}(model, order = "decreasing",
  sd_multi = 2, keep_intercept = FALSE, palette = "bilbao",
  ref_line = 0, trans = NULL, plot = TRUE, ranef = FALSE,
  which_ranef = NULL, ...)
}
\arguments{
\item{model}{The model that is the point of this function. For example, lm, glm, gam, lme4, brms.}

\item{order}{The order of the plots- "increasing", "decreasing", or a numeric
vector giving the order. The default is NULL, i.e. the default ordering. Not applied to random effects.}

\item{sd_multi}{The multiplier that determines the width of the interval. Default is 2.}

\item{keep_intercept}{Default is FALSE. Intercepts are typically on a very
different scale than covariate effects.}

\item{palette}{A scico palette. Default is 'bilbao'.}

\item{ref_line}{A reference line. Default is zero.}

\item{trans}{A transformation function to be applied to the coefficients
(e.g. exponentiation).}

\item{plot}{Default is TRUE, but sometimes you just want the data.}

\item{ranef}{If applicable, whether to plot random effects instead of fixed effects.}

\item{which_ranef}{If plotting random effects, which one to plot.}

\item{...}{Other arguments applied for specific methods.}
}
\value{
A ggplot of the coefficients and their interval estimates. Or the
  data that would be used to create the plot.
}
\description{
Plot fixed or random effects coefficients for merMod objects.
}
\details{
This plots the fixed or random effects of lme4 objects. For more
  information on the fixed effects, see \link[visibly]{plot_coefficients}.It
  requires the \link[lme4]{lme4} package.  The plot for random effects is
  basically the dotplot demonstrated at \code{?lme4::ranef}, but instead uses
  \link[ggplot2]{ggplot2} so you would have a little easier time working with
  it to do with as you wish (for multiple random effects, a list of ggplot
  objects can be returned). Many of the options for fixed effects are
  removed, as they either don't make much sense or for practical reasons.
}
\examples{
library(lme4)
fit_mer = lmer(Reaction ~ Days + (Days|Subject), sleepstudy)
plot_coefficients(fit_mer, ranef = TRUE, which_ranef = 'Subject')

}
\seealso{
Other model visualization: \code{\link{plot_coefficients.brmsfit}},
  \code{\link{plot_coefficients.lm}},
  \code{\link{plot_coefficients}},
  \code{\link{plot_gam_2d}}, \code{\link{plot_gam_3d}},
  \code{\link{plot_gam_check}}, \code{\link{plot_gam}}
}
\concept{model visualization}
