% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coefficients.R
\name{plot_coefficients}
\alias{plot_coefficients}
\title{Plot coefficients}
\usage{
plot_coefficients(model, ...)
}
\arguments{
\item{model}{The model. For example, lm, glm, gam, lme4, brms.}

\item{...}{Other arguments applied for specific methods.}
}
\value{
A ggplot of the coefficients and their interval estimates. Or the
data that would be used to create the plot.
}
\description{
A basic plot of coefficients with their uncertainty interval.
}
\examples{
mod = lm(mpg ~ ., mtcars)
plot_coefficients(mod, order = 'increasing')

}
\seealso{
Other model visualization: 
\code{\link{plot_coefficients.brmsfit}()},
\code{\link{plot_coefficients.lm}()},
\code{\link{plot_coefficients.merMod}()},
\code{\link{plot_gam_2d}()},
\code{\link{plot_gam_3d}()},
\code{\link{plot_gam_check}()},
\code{\link{plot_gam}()}
}
\concept{model visualization}
