% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coefs.R
\name{plot_coefs}
\alias{plot_coefs}
\title{Plot coefficients with uncertainty}
\usage{
plot_coefs(model_input, palette, ref_line, trans)
}
\arguments{
\item{model_input}{Processed model effects}

\item{palette}{Passed from other methods.}

\item{ref_line}{Passed from other methods.}

\item{trans}{Passed from other methods.}
}
\value{
A ggplot of the coefficients with their corresponding uncertainty
bars.
}
\description{
This isn't meant to be directly called, but is instead
internally used by the plot_coefficients function.
}
