% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fixefs.R
\name{plot_fixefs}
\alias{plot_fixefs}
\alias{plot_fixefs.brmsfit}
\alias{plot_fixefs.merMod}
\title{Plot coefficients with uncertainty}
\usage{
plot_fixefs(model, order = "decreasing", sd_multi = 2,
  keep_intercept = FALSE, palette = "bilbao", ref_line = 0,
  trans = NULL, plot = TRUE, ...)

\method{plot_fixefs}{brmsfit}(model, order, sd_multi, keep_intercept,
  palette, ref_line, trans, plot, ...)

\method{plot_fixefs}{merMod}(model, order, sd_multi, keep_intercept,
  palette, ref_line, trans, plot, ...)
}
\arguments{
\item{model}{The model that is the point of this function. For example, lm, glm, gam, lme4, brms.}

\item{order}{The order of the plots- "increasing", "decreasing", or a numeric
vector giving the order. The default is NULL, i.e. the default ordering. Not applied to random effects.}

\item{sd_multi}{The multiplier that determines the width of the interval. Default is 2.}

\item{keep_intercept}{Default is FALSE. Intercepts are typically on a very
different scale than covariate effects.}

\item{palette}{A scico palette. Default is 'bilbao'.}

\item{ref_line}{A reference line. Default is zero.}

\item{trans}{A transformation function to be applied to the coefficients
(e.g. exponentiation).}

\item{plot}{Default is TRUE, but sometimes you just want the data.}

\item{...}{Other arguments applied for specific methods.}
}
\value{
a ggplot2 object or the effect estimates
}
\description{
This isn't really meant to be directly called, but is instead
  internally used by the plot_coefficients function.
}
\examples{
#placeholder
}
\seealso{
\link[visibly]{plot_coefficients}
}
