% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gam_2d.R
\name{plot_gam_2d}
\alias{plot_gam_2d}
\alias{plot_gam_by}
\title{Plot 2-way GAM smooths
Plot 2-dimensional smooth terms}
\usage{
plot_gam_2d(model, main_var, second_var, conditional_data = NULL,
  n_plot = 100, force_2d = FALSE, ...)

plot_gam_by(model, main_var, by_var, conditional_data = NULL, n_plot = 500,
  ...)
}
\arguments{
\item{model}{The mgcv GAM.}

\item{main_var}{Which variable do you want to plot? Uses bare variable names
and can take multiple variables via \code{vars()}.}

\item{second_var}{Required for plot_gam_2d. The second continuous variable of
interest.}

\item{conditional_data}{This is the same as the newdata argument for predict.
Supply a data frame with desired values of the model covariates.}

\item{n_plot}{How many plotting points for the main_var/second_var? Default
is 100, creating a 100 x 100 grid of points.}

\item{force_2d}{If the second_var has <= 5 values, the plot_gam_by is called.
This will override that.}

\item{...}{Options to scale_fill_viridis_c for plot_gam_2d or
scale_color_viridis_d for plot_gam_by}

\item{by_var}{Required for plot_gam_by. The categorical variable of interest.}
}
\value{
A ggplot of the 2d effect.
}
\description{
Plot 2-way GAM smooths
Plot 2-dimensional smooth terms
}
\details{
These functions plot the predictions for two covariates in a GAM
  model produced by the \link[mgcv]{mgcv} package. The \code{plot_gam_2d}
  function is used for plotting two continuous predictors, while
  \code{plot_gam_by} is used in the case where one of the variables is
  categorical. If \code{plot_gam_2d} is called with the second variable being
  categorical or of very few distinct values, a message will follow along
  with a switch to \code{plot_gam_by}. One can override this with the
  \code{force_2d} argument.
}
\note{
Any attempt to use a non-numeric variable for the main_var will result in
  failure.

If you are using gamm or gamm4 then you need to supply the mgcv model
  as the model object.
}
\examples{
library(mgcv); library(dplyr)
set.seed(0)

d = gamSim(2, scale=.1)$data
mod <- gam(y ~ s(x, z), data = d)
plot_gam_2d(mod, main_var = x, second_var = z)
plot_gam_2d(mod, main_var = x, second_var = z, option='C')

d2 = gamSim(4)
mod_by <- gam(y ~ s(x2, by=fac), data = d2)
plot_gam_by(mod_by, main_var = x2, by_var = fac)

}
\seealso{
Other model visualization: \code{\link{plot_coefficients.brmsfit}},
  \code{\link{plot_coefficients.lm}},
  \code{\link{plot_coefficients.merMod}},
  \code{\link{plot_coefficients}},
  \code{\link{plot_gam_check}}, \code{\link{plot_gam}}
}
