% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gam_3d.R
\name{plot_gam_3d}
\alias{plot_gam_3d}
\title{Plot 2d smooths in 3d}
\usage{
plot_gam_3d(
  model,
  main_var,
  second_var,
  conditional_data = NULL,
  n_plot = 100,
  dmb = FALSE,
  ...
)
}
\arguments{
\item{model}{The mgcv gam model}

\item{main_var}{The 'x' axis.}

\item{second_var}{The 'y' axis'}

\item{conditional_data}{Values for other covariates. Default is NULL see
details.}

\item{n_plot}{Points to plot.  100 (the default) works well.  Embiggen at the
cost of your own waiting time.}

\item{dmb}{Whether to use plotly's display mode bar. Default is FALSE.}

\item{...}{Arguments for \link[scico]{scico}}
}
\value{
A plotly surface object
}
\description{
3d plot of 2d smooths for generalized additive models.
}
\details{
This works like \link[visibly]{plot_gam_2d}, the only difference
  being that a 3d plot is generated instead. It uses \link[scico]{scico} for
  the palette.  It is expected that the two input variables are continuous
}
\examples{
library(mgcv); library(visibly)
set.seed(0)

d = gamSim(2, scale=.1)$data
mod <- gam(y ~ s(x, z), data = d)
plot_gam_3d(mod, main_var = x, second_var = z)
plot_gam_3d(mod, main_var = x, second_var = z, palette='tokyo')
}
\seealso{
Other model visualization: 
\code{\link{plot_coefficients.brmsfit}()},
\code{\link{plot_coefficients.lm}()},
\code{\link{plot_coefficients.merMod}()},
\code{\link{plot_coefficients}()},
\code{\link{plot_gam_2d}()},
\code{\link{plot_gam_check}()},
\code{\link{plot_gam}()}
}
\concept{model visualization}
