% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gam_check.R
\name{plot_gam_check}
\alias{plot_gam_check}
\title{The visualization part of gam.check}
\usage{
plot_gam_check(model, single_page = TRUE, type = "deviance",
  scatter = FALSE, kcheck = FALSE)
}
\arguments{
\item{model}{The mgcv gam model}

\item{single_page}{Plot all on a single page. Requires
\link[gridExtra]{grid.arrange}}

\item{type}{The type of residuals wanted. Usually one of "deviance",
"pearson","scaled.pearson", "working", or "response"., See
\link[mgcv]{residuals.gam}.}

\item{scatter}{Whether to plot fitted vs. observed target variable as scatter
plot or density. Default is FALSE (density).}

\item{kcheck}{If you want the slightly less verbose basis dimension (k)
checking results. Default is FALSE. When true, only uses the defaults of
gam.check.  If you need to do more, use mgcv.}
}
\value{
A ggplot that provides visual inspection of residuals and more.
}
\description{
Residual plots etc.
}
\details{
Just a single page version of gam.check.
}
\examples{
library(mgcv); library(visibly)

d <- gamSim(1, n = 400)

g_fit <- gam(y ~ s(x0) + s(x1) + s(x2) + s(x3), data = d)

plot_gam_check(g_fit)
plot_gam_check(g_fit, scatter=TRUE)

}
\seealso{
\link[mgcv]{gam.check}
}
