% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ranefs.R
\name{plot_ranefs}
\alias{plot_ranefs}
\alias{plot_ranefs.brmsfit}
\alias{plot_ranefs.merMod}
\title{Plot coefficients with uncertainty}
\usage{
plot_ranefs(model, sd_multi = 2, ref_line = 0, trans = NULL,
  plot = TRUE, which_ranef = NULL, ...)

\method{plot_ranefs}{brmsfit}(model, sd_multi, ref_line, trans, plot,
  which_ranef, ...)

\method{plot_ranefs}{merMod}(model, sd_multi, ref_line, trans, plot,
  which_ranef, ...)
}
\arguments{
\item{model}{The brmsfit or lme4 model}

\item{sd_multi}{The sd multiplier}

\item{ref_line}{The reference line}

\item{trans}{A transformation to be applied to the coefficient.  Currently unused.}

\item{plot}{Whether to plot or just provide the data.}

\item{which_ranef}{Which random effect to plot}

\item{...}{Other options passed to specific methods.  Currently unused.}
}
\value{
a ggplot2 object or the effect estimates
}
\description{
This isn't really meant to be directly called, but is instead
  internally used by the plot_coefficients function.
}
\section{Methods (by class)}{
\itemize{
\item \code{brmsfit}: 

\item \code{merMod}: 
}}

\examples{
#placeholder
}
\seealso{
\link[visibly]{plot_coefficients}
}
