% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visibly.R
\docType{package}
\name{visibly}
\alias{visibly}
\title{visibly: Functions and palettes related to visual enhancement.}
\description{
Visibly is a handful of functions for color palettes, themes, etc. in R.  Inside you will find:
}
\details{
\itemize{
\item some ready-made palettes, e.g. based on R blue and Stan red
\item a function to quickly and easily create palettes with using \code{colortools::complementary} \code{colortools::adjacent} etc.
\item clean, web-friendly themes for ggplot2 and plotly
\item a function to interact with colorgorical
\item a heatmap for correlation matrices
\item coefficient plots for fixed and random effects, plotting of GAM results...
}

and more!
}
