% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coefs.R
\name{plot_coefs}
\alias{plot_coefs}
\title{Plot coefficients with uncertainty}
\usage{
plot_coefs(model_input, palette, ref_line, trans)
}
\arguments{
\item{model_input}{Processed model effects}

\item{palette}{A scico palette. Default is 'bilbao'.}

\item{ref_line}{A reference line. Default is zero.}

\item{trans}{A transformation function to be applied to the coefficients
(e.g. exponentiation).}
}
\value{
A ggplot of the coefficients with their corresponding uncertainty
  bars.
}
\description{
This isn't really meant to be directly called, but is instead
  internally used by the plot_coefficients function.
}
