% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{vars}
\alias{reexports}
\alias{map}
\alias{reexports}
\alias{aes}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{vars}}}

  \item{ggplot2}{\code{\link[ggplot2]{aes}}}

  \item{purrr}{\code{\link[purrr]{map}}}
}}

