% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_clean.R
\name{theme_trueMinimal}
\alias{theme_trueMinimal}
\alias{theme_plotly}
\alias{theme_blank}
\title{Clean visualization themes}
\usage{
theme_trueMinimal(font_size = 12, font_family = "",
  center_axis_labels = FALSE)

theme_plotly(vis, MB = FALSE)

theme_blank(vis, MB = FALSE)
}
\arguments{
\item{font_size}{Font size for axis labels in theme_trueMinimal}

\item{font_family}{Font family for axis labels in theme_trueMinimal}

\item{center_axis_labels}{Logical. Center axis labels in theme_trueMinimal.
Default is FALSE.}

\item{vis}{A visualization created by \link[ggplot2]{ggplot2} or
\link[plotly]{plotly}.}

\item{MB}{For plotly, an option to display the mode bar. Defaults to FALSE.}
}
\description{
Clean up plots from their defaults.
}
\details{
From a gray background, to unnecessary gridlines, to by-default
  reference lines, some of the more popular visualization packages come out
  75\% fantastic and 25\% questionable/arbitrary/problematic.  These functions
  remove unnecessary gridlines, 'de-bold' the blacks, etc.

  - \bold{ggplot2}: \code{theme_trueMinimal} This function takes a ggplot object
  and removes the gray background, gridlines and adds opacity to the default
  black axes and labels, allowing the pattern of the visual to be expressed
  in unimpeded fashion.

  - \bold{plotly}: \code{theme_plotly}, \code{theme_blank} removes reference
  lines at zero, and some of its 'modebar' is unnecessary.  Otherwise little
  is changed at this point, except for theme_blank, which is like theme_void
  for ggplot.
}
\note{
You may continue to override any aspect of these themes. For example with
ggplot2, you would just add a theme afterward just like you would any other
plot.
}
\examples{
library(visibly)
library(ggplot2)

data(mtcars)

ggplot(aes(wt, mpg), data=mtcars) +
  geom_point() +
  labs(title='Plot') +
  theme_trueMinimal()

library(plotly)
mtcars \%>\%
  plot_ly(x=~wt, y=~mpg, mode='markers') \%>\%
  theme_plotly()

mtcars \%>\%
  plot_ly(x=~wt, y=~mpg, mode='markers') \%>\%
  theme_blank()

}
\concept{themes}
